Fix a bug where restarting the container would ask for re-registration

+++ script/entrypoint.sh
@@ -1,11 +1,13 @@
 #!#!/bin/bash
 
-(echo ${ACCESS_CODE}; echo ${SERIAL_NUMBER}; echo "y" ; echo ${PRINTER_IP_ADDRESS}) |\
- ${OE_ENV}/bin/python3 -B -m py_installer ${PY_LAUNCH_JSON} || true
+
+if [ ! -f /root/start.sh ]; then
+    (echo ${ACCESS_CODE}; echo ${SERIAL_NUMBER}; echo "y" ; echo ${PRINTER_IP_ADDRESS}) |\
+    ${OE_ENV}/bin/python3 -B -m py_installer ${PY_LAUNCH_JSON} || true
 
 
-# Code taken from https://github.com/RyanEwen/octoeverywhere_bambu_docker/blob/main/scripts/container_setup.sh
- grep 'ExecStart=' /etc/systemd/system/octoeverywhere-bambu.service | \
- sed -r 's/ExecStart=//' > /root/start.sh && chmod +x /root/start.sh
-
- /root/start.sh
+    # Code taken from https://github.com/RyanEwen/octoeverywhere_bambu_docker/blob/main/scripts/container_setup.sh
+    grep 'ExecStart=' /etc/systemd/system/octoeverywhere-bambu.service | \
+    sed -r 's/ExecStart=//' > /root/start.sh && chmod +x /root/start.sh
+fi
+/root/start.sh
